src/requirements.txt
dash
pandas
plotly
psycopg2
SQLAlchemy
gunicorn
python-dotenv
openpyxl
